(PC-35769)[PRO] feat: refactoring cinema flow
