--- pro/src/pages/VenueSettings/VenueProvidersManager/AllocineProviderForm/AllocineProviderForm.module.scss
@@ -1,27 +1,0 @@
-@use "styles/mixins/_rem.scss" as rem;
-
-.form-content {
-  flex-grow: 1;
-}
-
-.form-layout-row {
-  display: flex;
-  flex-wrap: wrap;
-  align-items: flex-end;
-  gap: rem.torem(24px);
-  margin-bottom: rem.torem(16px);
-}
-
-.price-input,
-.nb-places-input {
-  width: rem.torem(220px);
-}
-
-.allocine-provider-form-banner {
-  margin: rem.torem(32px) 0;
-}
-
-.allocine-provider-form-actions {
-  display: flex;
-  gap: rem.torem(16px);
-}
