--- pro/src/pages/VenueSettings/VenueProvidersManager/CinemaProviderForm/constants.ts
@@ -1,6 +1,0 @@
-import { CinemaProviderFormValues } from './types'
-
-export const DEFAULT_CINEMA_PROVIDER_FORM_VALUES: CinemaProviderFormValues = {
-  isDuo: true,
-  isActive: true,
-}
