--- pro/src/pages/VenueSettings/VenueProvidersManager/CinemaProviderForm/types.ts
@@ -1,4 +1,0 @@
-export interface CinemaProviderFormValues {
-  isDuo?: boolean
-  isActive?: boolean
-}
