--- pro/src/pages/VenueSettings/VenueProvidersManager/DuoCheckbox/DuoCheckbox.module.scss
@@ -1,37 +1,0 @@
-@use "styles/variables/_forms.scss" as forms;
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_outline.scss" as outline;
-
-.duo-checkbox {
-  display: flex;
-  align-items: center;
-  gap: rem.torem(12px);
-  padding: rem.torem(16px);
-  border: rem.torem(1px) solid var(--color-grey-dark);
-  border-radius: rem.torem(8px);
-  margin-bottom: rem.torem(32px);
-
-  &:hover {
-    box-shadow: forms.$input-hover-shadow;
-    cursor: pointer;
-  }
-
-  &:focus-within {
-    @include outline.focus-outline(rem.torem(4px));
-  }
-
-  &-selected {
-    @include fonts.body-accent;
-
-    border: rem.torem(2px) solid var(--color-border-brand-primary);
-    background-color: var(--color-background-info);
-  }
-
-  &-icon {
-    flex-shrink: 0;
-    width: rem.torem(40px);
-    height: rem.torem(40px);
-    color: var(--color-icon-brand-primary);
-  }
-}
