--- pro/src/pages/VenueSettings/VenueProvidersManager/VenueProviderList/AllocineProviderEdit.module.scss
@@ -1,13 +1,0 @@
-@use "styles/mixins/_rem.scss" as rem;
-@use "styles/mixins/_fonts.scss" as fonts;
-@use "styles/mixins/_size.scss" as size;
-
-.allocine-provider-form {
-  display: flex;
-  flex-direction: column;
-  flex-grow: 1;
-}
-
-.explanation {
-  margin: rem.torem(8px) 0 rem.torem(32px) 0;
-}
